batch: Extract ownership absence claims
AbsenceClaim still lived in the aggregate ownership module even after
presence claims, references, replacement units, and metadata loading moved
to focused modules.

That made ownership.py act as both aggregate model and deletion-claim
record owner, and kept callers tied to the old module for absence values.

This commit moves AbsenceClaim into ownership_absence_claims and updates
runtime imports to use that value module directly.

BatchOwnership now imports the type privately for annotations, avoiding a
public old-location alias while keeping ownership aggregation unchanged.
