+++ src/git_stage_batch/batch/absence_constraints.py
@@ -28,7 +28,7 @@ from .realized_entry_storage import (
 )
 
 if TYPE_CHECKING:
-    from .ownership import AbsenceClaim
+    from .ownership_absence_claims import AbsenceClaim
 
 
 _DEFAULT_CHOICE_SCAN_CAP = 50
