+++ src/git_stage_batch/batch/baseline_replacement_choices.py
@@ -11,7 +11,7 @@ from ..core.text_lines import normalize_line_endings
 from .line_sequence_equality import line_slice_equals as _line_slice_matches
 
 if TYPE_CHECKING:
-    from .ownership import AbsenceClaim
+    from .ownership_absence_claims import AbsenceClaim
 
 
 @dataclass(frozen=True)
