+++ src/git_stage_batch/batch/hunk_ownership_translation.py
@@ -10,10 +10,8 @@ from .absence_content import (
     AbsenceContentBuilder as _AbsenceContentBuilder,
     build_absence_content_from_range as _build_absence_content_from_range,
 )
-from .ownership import (
-    AbsenceClaim,
-    BatchOwnership,
-)
+from .ownership import BatchOwnership
+from .ownership_absence_claims import AbsenceClaim
 from .ownership_claims import LineRangeBuilder, presence_claims_from_source_lines
 from .ownership_line_entries import (
     LineEntryContentSequence as _LineEntryContentSequence,
