+++ src/git_stage_batch/batch/ownership_merging.py
@@ -6,10 +6,8 @@ from dataclasses import dataclass
 from hashlib import sha256
 
 from ..core.buffer import buffer_byte_chunks
-from .ownership import (
-    AbsenceClaim,
-    BatchOwnership,
-)
+from .ownership import BatchOwnership
+from .ownership_absence_claims import AbsenceClaim
 from .ownership_claims import presence_claims_from_source_lines
 from .ownership_references import BaselineReference
 from .ownership_replacement_units import ReplacementUnit, normalize_replacement_units
