+++ src/git_stage_batch/batch/ownership_remapping.py
@@ -8,10 +8,8 @@ from ..core.line_selection import LineSelection
 from .lineage import BatchSourceLineage
 from .line_mapping import LineMapping
 from .match import match_lines
-from .ownership import (
-    AbsenceClaim,
-    BatchOwnership,
-)
+from .ownership import BatchOwnership
+from .ownership_absence_claims import AbsenceClaim
 from .ownership_claims import (
     format_ownership_line_set,
     parse_ownership_line_ranges,
