+++ src/git_stage_batch/batch/presence_constraints.py
@@ -28,7 +28,7 @@ from ..exceptions import MergeError as _MergeError
 from ..i18n import _
 
 if TYPE_CHECKING:
-    from .ownership import AbsenceClaim
+    from .ownership_absence_claims import AbsenceClaim
 
 
 _PRESENCE_CANDIDATE_CAP = 50
