+++ src/git_stage_batch/batch/replacement.py
@@ -12,7 +12,8 @@ from ..core.replacement import (
     replacement_line_chunks,
 )
 from ..core.buffer import LineBuffer
-from .ownership import BatchOwnership, AbsenceClaim
+from .ownership import BatchOwnership
+from .ownership_absence_claims import AbsenceClaim
 from .ownership_replacement_units import ReplacementUnit
