+++ src/git_stage_batch/commands/batch_transform/sift_results.py
@@ -13,7 +13,8 @@ from ...batch.comparison import (
 )
 from ...batch.merge import merge_batch_from_line_sequences_as_buffer
 from ...batch.absence_content import AbsenceContentBuilder
-from ...batch.ownership import BatchOwnership, AbsenceClaim
+from ...batch.ownership import BatchOwnership
+from ...batch.ownership_absence_claims import AbsenceClaim
 from ...batch.ownership_metadata_loading import acquire_ownership_for_metadata_dict
 from ...batch.query import get_batch_baseline_commit
 from ...batch.realized_file_content import build_realized_buffer_from_lines
